+++ apps/nextjs/.storybook/preview.tsx
@@ -8,12 +8,17 @@ import "@fontsource/lexend/800.css";
 import "@fontsource/lexend/900.css";
 import { OakThemeProvider, oakDefaultTheme } from "@oaknational/oak-components";
 import type { Preview, Decorator } from "@storybook/react";
+import { initialize as initializeMsw, mswLoader } from "msw-storybook-addon";
 
 import { TooltipProvider } from "../src/components/AppComponents/Chat/ui/tooltip";
+import { DialogProvider } from "../src/components/AppComponents/DialogContext";
 import { AnalyticsProvider } from "../src/mocks/analytics/provider";
+import { ClerkDecorator } from "../src/mocks/clerk/ClerkDecorator";
 import { TRPCReactProvider } from "../src/utils/trpc";
 import { RadixThemeDecorator } from "./decorators/RadixThemeDecorator";
 import "./preview.css";
+
+initializeMsw();
 
 const preview: Preview = {
   parameters: {
@@ -24,29 +29,32 @@ const preview: Preview = {
       },
     },
   },
-  tags: ["autodocs"],
+  loaders: [mswLoader],
 };
 
 // Providers not currently used
-// - MockClerkProvider
 // - CookieConsentProvider
 // - DemoProvider
 // - LessonPlanTrackingProvider
 // - DialogProvider
-// - OakThemeProvider
 // - SidebarProvider
 // - ChatModerationProvider
 
 export const decorators: Decorator[] = [
   RadixThemeDecorator,
+  ClerkDecorator,
   (Story) => (
     <>
       {/* TODO: Mock tRPC calls with MSW */}
       <TRPCReactProvider>
         <AnalyticsProvider>
-          <TooltipProvider>
-            <Story />
-          </TooltipProvider>
+          <DialogProvider>
+            <OakThemeProvider theme={oakDefaultTheme}>
+              <TooltipProvider>
+                <Story />
+              </TooltipProvider>
+            </OakThemeProvider>
+          </DialogProvider>
         </AnalyticsProvider>
       </TRPCReactProvider>
     </>