+++ apps/nextjs/src/components/AppComponents/Chat/chat-message/index.tsx
@@ -135,17 +135,17 @@ export function ChatMessage({
         errorType={hasError ? "generic" : null}
         type={getAvatarType()}
       >
-        <div
-          className="absolute left-0 top-0 h-20 w-20 "
+        <button
+          className="absolute left-0 top-0 h-20 w-20"
           onClick={() => {
             setInspect(!inspect);
           }}
         />
         <MessageTextWrapper>
           {message.id !== "working-on-it-initial" &&
-            messageParts.map((part, index) => {
+            messageParts.map((part) => {
               return (
-                <div className="w-full" key={index}>
+                <div className="w-full" key={part.id}>
                   <ChatMessagePart
                     part={part}
                     moderationModalHelpers={moderationModalHelpers}
@@ -238,6 +238,7 @@ function ChatMessagePart({
     error: ErrorMessagePart,
     bad: BadMessagePart,
     patch: PatchMessagePart,
+    experimentalPatch: ExperimentalPatchMessageComponent,
     state: StateMessagePart,
     text: TextMessagePart,
     action: ActionMessagePart,
@@ -295,7 +296,7 @@ function ErrorMessagePart({
 }: Readonly<{
   part: ErrorDocument;
 }>) {
-  const markdown = part.message || "Sorry, an error has occurred";
+  const markdown = part.message ?? "Sorry, an error has occurred";
   return <MemoizedReactMarkdownWithStyles markdown={markdown} />;
 }
 
@@ -337,4 +338,12 @@ function PartInspector({ part }: Readonly<{ part: MessagePart }>) {
       </pre>
     </div>
   );
+}
+
+/**
+ * Patches do not get rendered, they get applied to the lesson plan
+ * state, which is then rendered in the right hand side.
+ */
+function ExperimentalPatchMessageComponent() {
+  return null;
 }