+++ apps/nextjs/src/components/AppComponents/Chat/drop-down-section/chat-section.tsx
@@ -4,16 +4,15 @@ import { lessonSectionTitlesAndMiniDescriptions } from "data/lessonSectionTitles
 
 import { sectionTitle } from ".";
 import { MemoizedReactMarkdownWithStyles } from "../markdown";
+import AddAdditionalMaterialsButton from "./add-additional-materials-button";
 import FlagButton from "./flag-button";
 import ModifyButton from "./modify-button";
 
-const ChatSection = ({
-  objectKey,
-  value,
-}: {
+export type ChatSectionProps = Readonly<{
   objectKey: string;
   value: Record<string, unknown> | string | Array<unknown>;
-}) => {
+}>;
+const ChatSection = ({ objectKey, value }: ChatSectionProps) => {
   return (
     <OakFlex $flexDirection="column">
       <MemoizedReactMarkdownWithStyles
@@ -28,11 +27,20 @@ const ChatSection = ({
         $position="relative"
         $display={["none", "flex"]}
       >
-        <ModifyButton
-          sectionTitle={sectionTitle(objectKey)}
-          sectionPath={objectKey}
-          sectionValue={value}
-        />
+        {objectKey === "additionalMaterials" && value === "None" ? (
+          <AddAdditionalMaterialsButton
+            sectionTitle={sectionTitle(objectKey)}
+            sectionPath={objectKey}
+            sectionValue={value}
+          />
+        ) : (
+          <ModifyButton
+            sectionTitle={sectionTitle(objectKey)}
+            sectionPath={objectKey}
+            sectionValue={value}
+          />
+        )}
+
         <FlagButton
           sectionTitle={sectionTitle(objectKey)}
           sectionPath={objectKey}