+++ apps/nextjs/src/components/OakBoxWithCustomWidth.tsx
@@ -1,0 +1,8 @@
+import { OakBox } from "@oaknational/oak-components";
+import styled from "styled-components";
+
+const OakBoxWithCustomWidth = styled(OakBox)<{ width: number }>`
+  width: ${({ width }) => width}px;
+`;
+
+export default OakBoxWithCustomWidth;