+++ CHANGE_LOG.md
@@ -1,3 +1,45 @@
+## [1.16.2](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.16.1...v1.16.2) (2024-11-25)
+
+
+### Bug Fixes
+
+* add missing dependencies to Analytics Provider ([#306](https://github.com/oaknational/oak-ai-lesson-assistant/issues/306)) ([871c23f](https://github.com/oaknational/oak-ai-lesson-assistant/commit/871c23f4af0374d98428305fd388400ebb08b035))
+* handle aborts and add logging to the stream handler ([#350](https://github.com/oaknational/oak-ai-lesson-assistant/issues/350)) ([20f956e](https://github.com/oaknational/oak-ai-lesson-assistant/commit/20f956e35299ed8f465a8265d7aa52c7f18ac830))
+* hook error on sign-in page ([#208](https://github.com/oaknational/oak-ai-lesson-assistant/issues/208)) ([c298b10](https://github.com/oaknational/oak-ai-lesson-assistant/commit/c298b1051cab04643f04f6526b1dd0d835321c7e))
+* **sec:** bump next to 14.2.18 ([#242](https://github.com/oaknational/oak-ai-lesson-assistant/issues/242)) ([81fc31c](https://github.com/oaknational/oak-ai-lesson-assistant/commit/81fc31cddb3771b2ace860ee23b17448b22af7fb))
+
+## [1.16.1](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.16.0...v1.16.1) (2024-11-21)
+
+
+### Bug Fixes
+
+* add types for SVG imports, remove unused ([#318](https://github.com/oaknational/oak-ai-lesson-assistant/issues/318)) ([c1eff88](https://github.com/oaknational/oak-ai-lesson-assistant/commit/c1eff88efad2496ac0d82a8b8c03d38adf61e8ed))
+* streaming JSON types and error reporting tests ([#315](https://github.com/oaknational/oak-ai-lesson-assistant/issues/315)) ([3e18b44](https://github.com/oaknational/oak-ai-lesson-assistant/commit/3e18b446a58f44222919fa4712379785500b903d))
+
+# [1.16.0](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.15.0...v1.16.0) (2024-11-18)
+
+
+### Features
+
+* add FeatureFlagProvider ([#353](https://github.com/oaknational/oak-ai-lesson-assistant/issues/353)) ([1d4995e](https://github.com/oaknational/oak-ai-lesson-assistant/commit/1d4995ea0c82772259bc5312ba8d872dbd30b2b9))
+* link to hubspot form from requests for full access and higher rate - AI-626 AI-627 ([#359](https://github.com/oaknational/oak-ai-lesson-assistant/issues/359)) ([05ccce6](https://github.com/oaknational/oak-ai-lesson-assistant/commit/05ccce69348b03df2edee01dd1a27814a071be3d))
+
+# [1.15.0](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.14.2...v1.15.0) (2024-11-13)
+
+
+### Features
+
+* add additional materials button - AI-539 [migration] ([#255](https://github.com/oaknational/oak-ai-lesson-assistant/issues/255)) ([d0fe2d0](https://github.com/oaknational/oak-ai-lesson-assistant/commit/d0fe2d015865b89ea2287993652a6f8111f0ae4a))
+* prisma health check - AI-625 ([#356](https://github.com/oaknational/oak-ai-lesson-assistant/issues/356)) ([854950d](https://github.com/oaknational/oak-ai-lesson-assistant/commit/854950d51524eb8d84a0ec9695c88b67f829fd8d))
+
+## [1.14.2](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.14.1...v1.14.2) (2024-11-12)
+
+
+### Bug Fixes
+
+* design-changes-to-footer ([#324](https://github.com/oaknational/oak-ai-lesson-assistant/issues/324)) ([273cfdc](https://github.com/oaknational/oak-ai-lesson-assistant/commit/273cfdc668ca45def0b8a68dc08b7301974e1def))
+* only categorise initial user input once ([#348](https://github.com/oaknational/oak-ai-lesson-assistant/issues/348)) ([dd5bf71](https://github.com/oaknational/oak-ai-lesson-assistant/commit/dd5bf71a21421ac6e0beb60b4bab560cb159d877))
+
 ## [1.14.1](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.14.0...v1.14.1) (2024-11-07)